Renames Delimiter to HorizontalRule; Removes redundant attributes in comments style; Moves logic from quests-sets to quest-page. Also adds postcss-for plugin.
