+++ frontend/components/comments/styles.css
@@ -1,6 +1,4 @@
 .container {
-  display: flex;
-  flex-direction: column;
   border: 1px solid;
   margin: 5px;
 }
