+++ frontend/components/horizontal-rule/component.jsx
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+import styles from './styles.css';
+
+export default class HorizontalRule extends Component {
+  render() {
+    return (
+      <hr className={styles.horizontalRule} />
+    );
+  }
+}
