+++ frontend/components/horizontal-rule/index.js
