+++ frontend/components/horizontal-rule/styles.css
@@ -1,4 +1,4 @@
-hr.delimiter {
+hr.horizontal-rule {
   border-top: 2px solid gray;
   margin: 5px;
 }
