--- frontend/components/quests-sets/styles.css
@@ -1,7 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  border: 1px solid;
-  margin: 5px;
-}
